+++ LightBlue.Host/Properties/AssemblyInfo.cs
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("LightBlue.Host")]
-[assembly: AssemblyCopyright("Copyright © Colin David Scott 2014")]
+[assembly: AssemblyCopyright("Copyright © Colin David Scott 2014-2015")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -15,8 +15,8 @@ using System.Runtime.InteropServices;
 
 [assembly: Guid("619a7d7c-7ff5-4a23-80fb-110004e7ca0c")]
 
-[assembly: AssemblyVersion("1.1.16.0")]
-[assembly: AssemblyFileVersion("1.1.16.0")]
-[assembly: AssemblyInformationalVersion("1.1.16.0")]
+[assembly: AssemblyVersion("1.1.17.0")]
+[assembly: AssemblyFileVersion("1.1.17.0")]
+[assembly: AssemblyInformationalVersion("1.1.17.0")]
 
 [assembly: CLSCompliant(false)]