+++ LightBlue.Tests/Standalone/BlobStorage/StandaloneAzureBlobContainerListTests.cs
@@ -10,7 +10,7 @@ using Microsoft.WindowsAzure.Storage.Blob;
 using Xunit;
 using Xunit.Extensions;
 
-namespace LightBlue.Tests.Standalone
+namespace LightBlue.Tests.Standalone.BlobStorage
 {
     public class StandaloneAzureBlobContainerListTests : StandaloneAzureTestsBase, IDisposable
     {