+++ LightBlue.Tests/Standalone/BlobStorage/StandaloneAzureBlockBlobTests.cs
@@ -10,7 +10,7 @@ using LightBlue.Standalone;
 using Xunit;
 using Xunit.Extensions;
 
-namespace LightBlue.Tests.Standalone
+namespace LightBlue.Tests.Standalone.BlobStorage
 {
     public class StandaloneAzureBlockBlobTests : StandaloneAzureTestsBase
     {