+++ LightBlue/IAzureQueueStorageClient.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace LightBlue
+{
+    public interface IAzureQueueStorageClient
+    {
+        IAzureQueue GetQueueReference(string queueName);
+
+        IEnumerable<IAzureQueue> ListQueues();
+    }
+}