+++ LightBlue/Properties/AssemblyInfo.cs
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("LightBlue")]
-[assembly: AssemblyCopyright("Copyright © Colin David Scott 2014")]
+[assembly: AssemblyCopyright("Copyright © Colin David Scott 2014-2015")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -15,8 +15,8 @@ using System.Runtime.InteropServices;
 
 [assembly: Guid("743b834d-6da7-4cab-9199-4f80f9adf6a8")]
 
-[assembly: AssemblyVersion("1.1.16.0")]
-[assembly: AssemblyFileVersion("1.1.16.0")]
-[assembly: AssemblyInformationalVersion("1.1.16.0")]
+[assembly: AssemblyVersion("1.1.17.0")]
+[assembly: AssemblyFileVersion("1.1.17.0")]
+[assembly: AssemblyInformationalVersion("1.1.17.0")]
 
 [assembly: CLSCompliant(false)]