+++ TestWindsorRole/Properties/AssemblyInfo.cs
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("TestWindorRole")]
-[assembly: AssemblyCopyright("Copyright © Colin David Scott 2014")]
+[assembly: AssemblyCopyright("Copyright © Colin David Scott 2014-2015")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -15,8 +15,8 @@ using System.Runtime.InteropServices;
 
 [assembly: Guid("4aad25c2-57d6-43c5-a6da-f47246f1f68c")]
 
-[assembly: AssemblyVersion("1.1.16.0")]
-[assembly: AssemblyFileVersion("1.1.16.0")]
-[assembly: AssemblyInformationalVersion("1.1.16.0")]
+[assembly: AssemblyVersion("1.1.17.0")]
+[assembly: AssemblyFileVersion("1.1.17.0")]
+[assembly: AssemblyInformationalVersion("1.1.17.0")]
 
 [assembly: CLSCompliant(false)]